fixed generate recipe page
