--- src/pages/GenerateRecipe.jsx
@@ -1,5 +1,0 @@
-export default function GenerateRecipe () {
-    return (
-        <div>I am a form to generate a recipe.</div>
-    );
-}
