+++ src/pages/GenerateRecipe/GenerateRecipePage.jsx
@@ -1,0 +1,11 @@
+import GenerateRecipe from "../../components/recipes/GenerateRecipe";
+
+function GenerateRecipePage() {
+  return (
+    <div>
+      <GenerateRecipe />
+    </div>
+  );
+}
+
+export default GenerateRecipePage;
